requirements.txt
discord.py
pydantic
python-dotenv
aenum
pycountry
country_converter